ref(ui) Move EventsChart out of views
These components are used in a bunch of places and shouldn't be in the
views directory. Moving them will make events v1 easier to remove in the
future and better consolidates chart UI bits together.
