+++ src/sentry/static/sentry/app/actionCreators/events.tsx
@@ -3,7 +3,7 @@ import pick from 'lodash/pick';
 
 import {Client} from 'app/api';
 import {URL_PARAM} from 'app/constants/globalSelectionHeader';
-import {canIncludePreviousPeriod} from 'app/views/events/utils/canIncludePreviousPeriod';
+import {canIncludePreviousPeriod} from 'app/components/charts/utils';
 import {getPeriod} from 'app/utils/getPeriod';
 import {EventsStats, OrganizationSummary, MultiSeriesEventsStats} from 'app/types';
