+++ src/sentry/static/sentry/app/components/charts/eventsChart.jsx
@@ -18,7 +18,7 @@ import TransparentLoadingMask from 'app/components/charts/transparentLoadingMask
 import ErrorPanel from 'app/components/charts/errorPanel';
 import {getDuration, formatPercentage} from 'app/utils/formatters';
 
-import EventsRequest from './utils/eventsRequest';
+import EventsRequest from './eventsRequest';
 
 const DURATION_AGGREGATE_PATTERN = /^(p75|p95|p99|percentile)|transaction\.duration/;
 const PERCENTAGE_AGGREGATE_PATTERN = /^(error_rate)/;
