+++ src/sentry/static/sentry/app/components/charts/utils.tsx
@@ -83,3 +83,22 @@ export function getDiffInMinutes(datetimeObj: DateTimeObject): number {
     parsePeriodToHours(typeof period === 'string' ? period : DEFAULT_STATS_PERIOD) * 60
   );
 }
+
+// Max period (in hours) before we can no long include previous period
+const MAX_PERIOD_HOURS_INCLUDE_PREVIOUS = 45 * 24;
+
+export function canIncludePreviousPeriod(
+  includePrevious: boolean | undefined,
+  period: string | undefined
+) {
+  if (!includePrevious) {
+    return false;
+  }
+
+  if (period && parsePeriodToHours(period) > MAX_PERIOD_HOURS_INCLUDE_PREVIOUS) {
+    return false;
+  }
+
+  // otherwise true
+  return !!includePrevious;
+}
