--- src/sentry/static/sentry/app/views/events/utils/canIncludePreviousPeriod.jsx
@@ -1,16 +1,0 @@
-import {parsePeriodToHours} from 'app/utils/dates';
-// Max period (in hours) before we can no long include previous period
-const MAX_PERIOD_HOURS_INCLUDE_PREVIOUS = 45 * 24;
-
-export function canIncludePreviousPeriod(includePrevious, period) {
-  if (!includePrevious) {
-    return false;
-  }
-
-  if (period && parsePeriodToHours(period) > MAX_PERIOD_HOURS_INCLUDE_PREVIOUS) {
-    return false;
-  }
-
-  // otherwise true
-  return includePrevious;
-}
