+++ src/sentry/static/sentry/app/views/performance/transactionSummary/sidebarCharts.tsx
@@ -10,13 +10,13 @@ import EventView from 'app/utils/discover/eventView';
 import ChartZoom from 'app/components/charts/chartZoom';
 import LineChart from 'app/components/charts/lineChart';
 import ErrorPanel from 'app/components/charts/errorPanel';
+import EventsRequest from 'app/components/charts/eventsRequest';
 import QuestionTooltip from 'app/components/questionTooltip';
 import {SectionHeading} from 'app/components/charts/styles';
 import TransparentLoadingMask from 'app/components/charts/transparentLoadingMask';
 import TransitionChart from 'app/components/charts/transitionChart';
 import {getInterval} from 'app/components/charts/utils';
 import {IconWarning} from 'app/icons';
-import EventsRequest from 'app/views/events/utils/eventsRequest';
 import {PERFORMANCE_TERMS} from 'app/views/performance/constants';
 import {getUtcToLocalDateObject} from 'app/utils/dates';
 import {
