+++ src/sentry/static/sentry/app/views/settings/incidentRules/triggers/chart/index.tsx
@@ -5,7 +5,7 @@ import styled from '@emotion/styled';
 import {Client} from 'app/api';
 import {Organization, Project} from 'app/types';
 import {SeriesDataUnit} from 'app/types/echarts';
-import EventsRequest from 'app/views/events/utils/eventsRequest';
+import EventsRequest from 'app/components/charts/eventsRequest';
 import LoadingMask from 'app/components/loadingMask';
 import Placeholder from 'app/components/placeholder';
 import space from 'app/styles/space';
